Move all exceptions tests with run_test into a single file
