+++ plasmapy/particles/tests/test_nuclear.py
@@ -10,82 +10,7 @@ from plasmapy.particles.nuclear import (
     nuclear_binding_energy,
     nuclear_reaction_energy,
 )
-from plasmapy.utils.pytest_helpers import run_test, run_test_equivalent_calls
-
-test_nuclear_table = [
-    [nuclear_binding_energy, "p", {}, 0 * u.J],
-    [nuclear_binding_energy, "n", {}, 0 * u.J],
-    [nuclear_binding_energy, "p", {}, 0 * u.J],
-    [nuclear_binding_energy, "H", {}, ParticleError],
-    [nuclear_binding_energy, "He-99", {}, InvalidParticleError],
-    [nuclear_binding_energy, "He", {"mass_numb": 99}, InvalidParticleError],
-    [nuclear_binding_energy, 3.1415926535j, {}, TypeError],
-    [mass_energy, "e-", {}, (const.m_e * const.c ** 2).to(u.J)],
-    [mass_energy, "p+", {}, (const.m_p * const.c ** 2).to(u.J)],
-    [mass_energy, "H-1", {}, (const.m_p * const.c ** 2).to(u.J)],
-    [mass_energy, "H-1 0+", {}, (const.m_p * const.c ** 2).to(u.J)],
-    [mass_energy, "n", {}, (const.m_n * const.c ** 2).to(u.J)],
-    [nuclear_reaction_energy, (), {"reactants": ["n"], "products": 3}, TypeError],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["n"], "products": ["He-4"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["h"], "products": ["H-1"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["e-", "n"], "products": ["p+"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["e+", "n"], "products": ["p-"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["ksdf"], "products": ["H-3"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["H"], "products": ["H-1"]},
-        ParticleError,
-    ],
-    [
-        nuclear_reaction_energy,
-        (),
-        {"reactants": ["p"], "products": ["n", "n", "e-"]},
-        ParticleError,
-    ],
-    [nuclear_reaction_energy, "H + H --> H", {}, ParticleError],
-    [nuclear_reaction_energy, "H + H", {}, ParticleError],
-    [nuclear_reaction_energy, 1, {}, TypeError],
-    [nuclear_reaction_energy, "H-1 + H-1 --> H-1", {}, ParticleError],
-    [nuclear_reaction_energy, "p --> n", {}, ParticleError],
-    [
-        nuclear_reaction_energy,
-        "p --> p",
-        {"reactants": "p", "products": "p"},
-        ParticleError,
-    ],
-]
-
-
-@pytest.mark.parametrize("test_inputs", test_nuclear_table)
-def test_nuclear(test_inputs):
-    run_test(*test_inputs, rtol=1e-3)
-
+from plasmapy.utils.pytest_helpers import run_test_equivalent_calls
 
 test_nuclear_equivalent_calls_table = [
     [nuclear_binding_energy, ["He-4", {}], ["alpha", {}], ["He", {"mass_numb": 4}]]
